Modified Makefile to have separate conditional flags. Modified uart_puts to fix race condition bugs. Modified putc to handle interrupts. Updated circular queue datatypes from uint16_t to uint32_t. Updated uploader to have a fflush(stdout) right after OS loading.

+++ include/common/cq.h
@@ -9,14 +9,14 @@
 
 struct cq {
 	volatile uint8_t *buff;
-	volatile uint16_t start;
-	volatile uint16_t end;
-	volatile uint16_t curfreesize;
-	volatile uint16_t buffsize;
-	volatile uint16_t modsize;
+	volatile uint32_t start;
+	volatile uint32_t end;
+	volatile uint32_t curfreesize;
+	volatile uint32_t buffsize;
+	volatile uint32_t modsize;
 };
 
-void cq_init(struct cq * const cq,uint8_t *buff,uint16_t buffsize);
+void cq_init(struct cq * const cq,uint8_t *buff,uint32_t buffsize);
 uint8_t cq_is_empty(struct cq * const cq);
 uint8_t cq_is_full(struct cq * const cq);
 void cq_add(struct cq * const cq,uint8_t ch);
